requirements.txt
torch==2.2.0
torchvision==0.17.0
lightning==2.4.0
librosa==0.9.1
python_speech_features==0.6
opencv-python==4.10.0.84
scikit-video==1.1.11
pillow==11.3.0
dlib==19.24.9
numpy==1.26.4
pandas==2.1.1
scikit-learn==1.3.2
tqdm==4.65.2
PyYAML==6.0.2